+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Merged Quick Power, Quick Volume, and Rapidmenu
 - Support testkit firmware 3.60
 - Added "Restart" text to poweroff button (20 languages)
+- Configurable Quick Menu PS button push time
 
 ### Changelog - Quick Power